Move init labels to init pipeline module
